+++ classes/kohana/profilertoolbar.php
@@ -7,7 +7,7 @@
 */
 class Kohana_ProfilerToolbar {
 
-  public  static $version = '0.2.0';
+  public  static $version = '0.2.1';
   public  static $kohana_version = '3.2';
   private static $_cfg = null;
   /* @var FirePHP */
@@ -438,35 +438,155 @@ class Kohana_ProfilerToolbar {
     }
   }
 
-  public static function highlightSQL($sql = ''){
-    $words = array (
-      'chars' => '/([\\.,\\(\\)<>:=`]+)/i',
-      'constants' => '/(\'[^\']*\'|[0-9]+)/i',
-      'keywords' => array('SELECT', 'UPDATE', 'INSERT', 'DELETE', 'REPLACE', 'INTO', 'CREATE', 'ALTER', 'TABLE', 'DROP', 'TRUNCATE', 'FROM','ADD', 'CHANGE', 'COLUMN', 'KEY','WHERE', 'ON', 'CASE', 'WHEN', 'THEN', 'END', 'ELSE', 'AS','USING', 'USE', 'INDEX', 'CONSTRAINT', 'REFERENCES', 'DUPLICATE','LIMIT', 'OFFSET', 'SET', 'SHOW', 'STATUS','BETWEEN', 'AND', 'IS', 'NOT', 'OR', 'XOR', 'INTERVAL', 'TOP','GROUP BY', 'ORDER BY', 'DESC', 'ASC', 'COLLATE', 'NAMES', 'UTF8', 'DISTINCT', 'DATABASE','CALC_FOUND_ROWS', 'SQL_NO_CACHE', 'MATCH', 'AGAINST', 'LIKE', 'REGEXP', 'RLIKE','PRIMARY', 'AUTO_INCREMENT', 'DEFAULT', 'IDENTITY', 'VALUES', 'PROCEDURE', 'FUNCTION','TRAN', 'TRANSACTION', 'COMMIT', 'ROLLBACK', 'SAVEPOINT', 'TRIGGER', 'CASCADE','DECLARE', 'CURSOR', 'FOR', 'DEALLOCATE'),
-      'joins' => array('JOIN', 'INNER', 'OUTER', 'FULL', 'NATURAL', 'LEFT', 'RIGHT'),
-      'functions' => array('MIN', 'MAX', 'SUM', 'COUNT', 'AVG', 'CAST', 'COALESCE', 'CHAR_LENGTH', 'LENGTH', 'SUBSTRING','DAY', 'MONTH', 'YEAR', 'DATE_FORMAT', 'CRC32', 'CURDATE', 'SYSDATE', 'NOW', 'GETDATE','FROM_UNIXTIME', 'FROM_DAYS', 'TO_DAYS', 'HOUR', 'IFNULL', 'ISNULL', 'NVL', 'NVL2','INET_ATON', 'INET_NTOA', 'INSTR', 'FOUND_ROWS','LAST_INSERT_ID', 'LCASE', 'LOWER', 'UCASE', 'UPPER','LPAD','RPAD','RTRIM','LTRIM','MD5','MINUTE', 'ROUND','SECOND', 'SHA1', 'STDDEV', 'STR_TO_DATE', 'WEEK',),
-    );
+  /**
+   * Highlite code
+   * @static
+   * @param $source
+   * @param $lang (sql|php)
+   * @return string
+   */
+  public static function highlight($source, $lang){
+    if(!in_array($lang,array('sql','php'))) return $source;
+    $geshi = new GeSHi($source,$lang);
+    $geshi->enable_classes();
+    $geshi->enable_keyword_links(false);
+    $res = $geshi->parse_code();
+    $res = str_replace('<pre class="'.$lang.'">','',$res);
+    $res = str_replace('</pre>','',$res);
+    $res = trim(rtrim($res),"\n");
+    return $res;
+  }
 
-    // set \n when needed
-//    $sql = str_replace("\n",' ',$sql);
-//    $nlAfter = array('FROM','WHERE','GROUP BY','LIMIT','ORDER BY','LEFT JOIN','INNER JOIN');
-//    $sql = preg_replace('/\\b('.join("|", $nlAfter).')\\b/i', "\n$1", $sql);
-//    $sql = preg_replace('/\s/','&nbsp;',$sql);
+  /**
+   * get part of file source by line
+   * and can highlite it
+   * @static
+   * @param $file
+   * @param $line_number
+   * @param int $padding
+   * @param bool $highlite
+   * @param string $lang
+   * @return string
+   */
+  public static function debugSource($file, $line_number, $padding = 5, $highlite = false, $lang = ''){
+    if(!$file OR !is_readable($file)) return '';
 
-    $sql = str_replace('\\\'', '\\&#039;', $sql);
-    foreach($words as $key => $items){
-      if(in_array($key, array('constants', 'chars'))){
-        $regexp = $items;
-      }else{
-        $regexp = '/\\b('.join("|", $items).')\\b/i';
+    $file = fopen($file, 'r');
+    $line = 0;
+    $range = array('start' => $line_number - $padding, 'end' => $line_number + $padding);
+    $format = '% '.strlen($range['end']).'d';
+
+    $source = '';
+    while(($row = fgets($file)) !== FALSE){
+      if(++$line > $range['end']) break;
+      if($line >= $range['start']) $source .= $row;
+    }
+    fclose($file);
+
+    if($highlite){
+      $source = self::highlight($source,$lang);
+    }
+    
+    $source = explode("\n",$source);
+
+    $line = $range['start'];
+    $result = '';
+    foreach($source as $row){
+      $hRow = ($line == $line_number)?' highlight':'';
+      if($line === $line_number){
+        self::addData($row,'sec');
       }
-      $sql = preg_replace($regexp, '<span class="'.$key.'">$1</span>', $sql);
+      $result .= '<span class="line'.$hRow.'"><span class="num">'.sprintf($format, $line).' </span>'.$row.'</span>';
+      $line++;
     }
 
-//    $sql = nl2br($sql);
+    return $result;
+  }
 
-//    var_dump($sql);die();
-    return $sql;
+  /**
+   * get source by line in all files of trace
+   * @static
+   * @param array|null $trace
+   * @param bool $highlight
+   * @param string $lang
+   * @return array
+   */
+  public static function debugTrace(array $trace = NULL, $highlight = false, $lang = ''){
+    if ($trace === NULL) $trace = debug_backtrace();
+    // Non-standard function calls
+    $statements = array('include', 'include_once', 'require', 'require_once');
+
+    $output = array();
+    foreach($trace as $step){
+      if(!isset($step['function'])) continue;
+      if(isset($step['file']) AND isset($step['line'])){
+        // Include the source of this step
+        $source = self::debugSource($step['file'], $step['line'],5,$highlight,$lang);
+      }
+
+      if(isset($step['file'])){
+        $file = $step['file'];
+
+        if(isset($step['line'])){
+          $line = $step['line'];
+        }
+      }
+
+      // function()
+      $function = $step['function'];
+
+      if(in_array($step['function'], $statements)){
+        if(empty($step['args'])) $args = array();
+        else $args = array($step['args'][0]);
+      }elseif(isset($step['args'])){
+        if(!function_exists($step['function']) OR strpos($step['function'], '{closure}') !== FALSE){
+          // Introspection on closures or language constructs in a stack trace is impossible
+          $params = NULL;
+        }else{
+          if(isset($step['class'])){
+            if(method_exists($step['class'], $step['function'])){
+              $reflection = new ReflectionMethod($step['class'], $step['function']);
+            }else{
+              $reflection = new ReflectionMethod($step['class'], '__call');
+            }
+          }else{
+            $reflection = new ReflectionFunction($step['function']);
+          }
+
+          // Get the function parameters
+          $params = $reflection->getParameters();
+        }
+
+        $args = array();
+
+        foreach($step['args'] as $i => $arg){
+          if(isset($params[$i])){
+            // Assign the argument by the parameter name
+            $args[$params[$i]->name] = $arg;
+          }else{
+            // Assign the argument by number
+            $args[$i] = $arg;
+          }
+        }
+      }
+
+      if(isset($step['class'])){
+        // Class->method() or Class::method()
+        $function = $step['class'].$step['type'].$step['function'];
+      }
+
+      $output[] = array(
+        'function' => $function,
+        'args' => isset($args)?$args:NULL,
+        'file' => isset($file)?$file:NULL,
+        'line' => isset($line)?$line:NULL,
+        'source' => isset($source)?$source:NULL,
+      );
+
+      unset($function, $args, $file, $line, $source);
+    }
+
+    return $output;
   }
 
 }