+++ views/kohana/error.php
@@ -2,30 +2,58 @@
 <?php echo "{$type} [ {$code} ]\n{$message}";?>
 <?php else:?>
 
+<?php $highlightSQL = ProfilerToolbar::cfg('errorPage.highlightSQL'); ;?>
+<?php $highlightPHP = ProfilerToolbar::cfg('errorPage.highlightPHP'); ;?>
+
 <?php $error_id = uniqid('error');?>
-  <?php echo View::factory('PTB/highlight.css');?>
   <style type="text/css">
     body{background-image:url(data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAQAAAAECAAAAACMmsGiAAAAGXRFWHRTb2Z0d2FyZQBBZG9iZSBJbWFnZVJlYWR5ccllPAAAABdJREFUCB1jUlGRkWH6CQRMf//y8oJZAFUaCmUwcfODAAAAAElFTkSuQmCC);}
     #ptb_err{font-size: 10pt; font-family:sans-serif; text-align: left; color: gray; margin-top: 30px;}
     #ptb_err a{color:gray;}
-    #ptb_err pre{overflow: auto; white-space: pre-wrap; font-size: 9pt; line-height: 12pt; margin: 4px 0; border-radius: 3px;}
-    #ptb_err code{}
-    #ptb_err pre span.line { display: block; }
-    #ptb_err pre span.highlight { background: #414040; }
-    #ptb_err pre span.line span.number { color: #666; }
+    #ptb_err pre{}
+    #ptb_err pre.source{overflow: auto; white-space: pre-wrap; font-size: 9pt; line-height: 12pt; margin: 4px 0; border-radius: 3px; padding: 4px 5px 4px 8px; background-color: /*#171717*/#242323; color: white;}
+    #ptb_err pre.source span.line { display: block; }
+    #ptb_err pre.source span.highlight { background: #414040; }
+    #ptb_err pre.source span.line span.num { color: #666; }
     #ptb_err .head{margin-bottom: 4px; background: #8C0B0B; color:white; padding: 10px; border-radius: 3px;}
     #ptb_err .head .type{font-weight: bold; margin: 0;}
     #ptb_err .head .message{margin: 3px 0 0 0;}
     #ptb_err .content{padding: 0 10px; margin: 10px 0 0 0;}
     #ptb_err .content .filePath{margin: 0;}
     #ptb_err .content .filePath .lineNum{}
-    #ptb_err .content pre.source{}
     #ptb_err .content ol.trace{padding: 0 0 0 10px;}
     #ptb_err .content ol.trace li{margin: 4px 0;}
     #ptb_err .content ol.trace li pre.source{display: none;}
     #ptb_err .content ol.trace li table.arguments{display: none; font-size: 9pt; border-collapse: collapse;padding: 0;width: 100%; margin: 5px 0;}
     #ptb_err .content ol.trace li table.arguments td {border: 1px solid #252424;color: gray;padding: 3px;text-align: left;vertical-align: top;}
     #ptb_err .content ol.trace li table.arguments td.name{text-align: center; width: 30px; font-size: 9pt;}
+    <!-- highlight -->
+    #ptb_err pre.sql .imp {font-weight: bold; color: red;}
+    #ptb_err pre.sql .kw1 {color: #388fff;}
+    #ptb_err pre.sql .co1 {color: #808080;}
+    #ptb_err pre.sql .co2 {color: #808080;}
+    #ptb_err pre.sql .coMULTI {color: #808080;}
+    #ptb_err pre.sql .es0 {color: #000099; font-weight: bold;}
+    #ptb_err pre.sql .br0 {}
+    #ptb_err pre.sql .st0 {color: #ff9933;}
+    #ptb_err pre.sql .nu0 {color: #66ff00;}
+
+    #ptb_err pre.php .imp {font-weight: bold; color: red;} 
+    #ptb_err pre.php .kw1 {color: #388fff;}
+    #ptb_err pre.php .kw2 {color: #388fff; font-weight: bold;}
+    #ptb_err pre.php .kw3 {color: #388fff;}
+    #ptb_err pre.php .co1 {color: #808080;}
+    #ptb_err pre.php .co2 {color: #808080;}
+    #ptb_err pre.php .coMULTI {color: #808080;}
+    #ptb_err pre.php .es0 {color: #ffdf08; font-weight: bold;}
+    #ptb_err pre.php .br0 {}
+    #ptb_err pre.php .st0 {color: #ff9933;}
+    #ptb_err pre.php .nu0 {color: #66ff00;}
+    #ptb_err pre.php .me1 {}
+    #ptb_err pre.php .me2 {}
+    #ptb_err pre.php .re0 {color: #66ff66;}
+    #ptb_err pre.php .re1 {color: #ff9933}
+    <!-- /highlight -->
   </style>
 
   <script type="text/javascript">
@@ -37,20 +65,6 @@
       elem.style.display = disp == 'block' ? 'none' : 'block';
       return false;
     }
-    <?php echo View::factory('PTB/highlight.js');?>
-
-    function init(){
-      var ptb_err = document.getElementById('ptb_err');
-      var sql_els = ptb_err.getElementsByClassName('sql');
-      for(var i=0;i<sql_els.length;i++) hljs.highlightBlock(sql_els[i],'  ');
-    }
-
-    if(document.addEventListener){
-      window.addEventListener('load',init,false);
-    }else if(document.attachEvent){
-      window.attachEvent("onload",init);
-    }
-
   </script>
 
   <div id="ptb_err">
@@ -65,7 +79,11 @@
         $message =  UTF8::substr($message,0,$start);
       ?>
         <?php echo $message;?>
-        <pre><code class="sql"><?php echo trim(rtrim($sql),"\n");?></code></pre>
+        <?php if($highlightSQL):?>
+          <pre class="source sql"><?php echo ProfilerToolbar::highlight($sql,'sql')?></pre>
+        <?php else:?>
+          <pre class="source"><?php echo $sql;?></pre>
+        <?php endif;?>
       <?php else:?>
       <?php echo html::chars($message); ?>
       <?php endif;?>
@@ -75,19 +93,14 @@
 
     <div class="content" id="<?php echo $error_id ?>">
       <p class="filePath"><?php echo Debug::path($file);?> <span class="lineNum">[ <?php echo $line;?> ]</span></p>
-      <pre class="source"><!--
-      --><code class="php"><!--
-        --><?php
-            $source = Debug::source($file, $line);
-            $source = UTF8::str_ireplace('<pre class="source"><code>','',$source);
-            $source = UTF8::str_ireplace('</code></pre>','',$source);
-            echo trim($source);
-          ?><!--
-      --></code><!--
-    --></pre>
+      <?php if($highlightPHP):?>
+        <pre class="source php"><?php echo ProfilerToolbar::debugSource($file, $line,5,true,'php');?></pre>
+      <?php else:?>
+        <pre class="source"><?php echo ProfilerToolbar::debugSource($file, $line);?></pre>
+      <?php endif;?>
 
       <ol class="trace">
-        <?php foreach(Debug::trace($trace) as $i => $step): ?>
+        <?php foreach(ProfilerToolbar::debugTrace($trace,$highlightPHP,'php') as $i => $step): ?>
         <li>
           <p class="filePath">
             <?php if($step['file']): $source_id = $error_id.'source'.$i; ?>
@@ -116,16 +129,11 @@
           <?php endif;?>
 
           <?php if(isset($source_id)): ?>
-          <pre class="source" id="<?php echo $source_id ?>"><!--
-          --><code class="php"><!--
-            --><?php
-                $source = $step['source'];
-                $source = UTF8::str_ireplace('<pre class="source"><code>','',$source);
-                $source = UTF8::str_ireplace('</code></pre>','',$source);
-                echo trim($source);
-              ?><!--
-          --></code><!--
-        --></pre>
+            <?php if($highlightPHP):?>
+            <pre class="source php" id="<?php echo $source_id ?>"><?php echo $step['source'];?></pre>
+            <?php else:?>
+            <pre class="source" id="<?php echo $source_id ?>"><?php echo $step['source'];?></pre>
+            <?php endif;?>
           <?php endif ?>
 
         </li>